Retrieve a data frame of all active surveys on Qualtrics

Retrieve a data frame of all active surveys on Qualtrics

all_surveys()

See also

Examples

# NOT RUN {
# Register your Qualtrics credentials if you haven't already
qualtrics_api_credentials(
  api_key = "<YOUR-API-KEY>",
  base_url = "<YOUR-BASE-URL>"
)

# Retrieve a list of all surveys
surveys <- all_surveys()

# Retrieve a single survey
mysurvey <- fetch_survey(surveyID = surveys$id[6])

# You can use the same parameters as those found in the Qualtrics API documentation
# Found here: https://api.qualtrics.com/docs/csv
mysurvey <- fetch_survey(
  surveyID = surveys$id[6],
  save_dir = tempdir(),
  start_date = "2018-01-01",
  end_date = "2018-01-31",
  limit = 100,
  label = TRUE,
  unanswer_recode = "UNANS",
  verbose = TRUE
)
# }