Read a CSV file exported from Qualtrics

Reads comma separated CSV files generated by Qualtrics software. The second line containing the variable labels is imported. Repetitive introductions to matrix questions are automatically removed. Variable labels are stored as attributes.

read_survey(file_name, strip_html = TRUE, legacy = FALSE)

Arguments

file_name
String. A CSV data file.

strip_html
Logical. If TRUE, then remove HTML tags. Defaults to TRUE.

legacy
Logical. If TRUE, then import "legacy" format CSV files (as of 2017). Defaults to FALSE.

Value

A data frame. Variable labels are stored as attributes. They are not printed on the console but are visibile in the RStudio viewer.

Examples

# NOT RUN {
# Generic use of read_survey()
df <- read_survey("<YOUR-PATH-TO-CSV-FILE>")
# }
# Example using current data format
file <- system.file("extdata", "sample.csv", package = "qualtRics")
df <- read_survey(file)

# Example using legacy data format
file <- system.file("extdata", "sample_legacy.csv", package = "qualtRics")
df <- read_survey(file, legacy = TRUE)

#> Warning: Missing column names filled in: 'X16' [16]